feat: MCCトレイト(MyCompiledContract)を実装した． ・CompiledContractトレイトに定義されていたmy_rsaメソッドもMCCに移植した． ・MCCトレイトを用いてLeviathanにmy_rsaを再実装した．
bench_runner, state_runner両テストでpassを確認
