+++ gas_analy/rsa_benchmarks.csv
@@ -4,3 +4,5 @@ PayloadLen:547,Status:Success,Time_us:175
 PayloadLen:547,Status:Success,Time_us:191
 PayloadLen:547,Status:Success,Time_us:194
 PayloadLen:547,Status:Success,Time_us:197
+PayloadLen:547,Status:Success,Time_us:154
+PayloadLen:547,Status:Success,Time_us:157
